+++ zerver/tests/test_subs.py
@@ -2995,7 +2995,7 @@ class StreamAdminTest(ZulipTestCase):
         If you're not an admin, you can't remove other people from streams except your own bots.
         """
         result = self.attempt_unsubscribe_of_principal(
-            query_count=8,
+            query_count=7,
             target_users=[self.example_user("cordelia")],
             is_realm_admin=False,
             is_subbed=True,
@@ -3010,7 +3010,7 @@ class StreamAdminTest(ZulipTestCase):
         those you aren't on.
         """
         result = self.attempt_unsubscribe_of_principal(
-            query_count=15,
+            query_count=14,
             target_users=[self.example_user("cordelia")],
             is_realm_admin=True,
             is_subbed=True,
@@ -3037,7 +3037,7 @@ class StreamAdminTest(ZulipTestCase):
             for name in ["cordelia", "prospero", "iago", "hamlet", "outgoing_webhook_bot"]
         ]
         result = self.attempt_unsubscribe_of_principal(
-            query_count=22,
+            query_count=21,
             cache_count=8,
             target_users=target_users,
             is_realm_admin=True,
@@ -3055,7 +3055,7 @@ class StreamAdminTest(ZulipTestCase):
         are on.
         """
         result = self.attempt_unsubscribe_of_principal(
-            query_count=15,
+            query_count=14,
             target_users=[self.example_user("cordelia")],
             is_realm_admin=True,
             is_subbed=True,
@@ -3072,7 +3072,7 @@ class StreamAdminTest(ZulipTestCase):
         streams you aren't on.
         """
         result = self.attempt_unsubscribe_of_principal(
-            query_count=15,
+            query_count=14,
             target_users=[self.example_user("cordelia")],
             is_realm_admin=True,
             is_subbed=False,
@@ -3086,7 +3086,7 @@ class StreamAdminTest(ZulipTestCase):
 
     def test_cant_remove_others_from_stream_legacy_emails(self) -> None:
         result = self.attempt_unsubscribe_of_principal(
-            query_count=8,
+            query_count=7,
             is_realm_admin=False,
             is_subbed=True,
             invite_only=False,
@@ -3098,7 +3098,7 @@ class StreamAdminTest(ZulipTestCase):
 
     def test_admin_remove_others_from_stream_legacy_emails(self) -> None:
         result = self.attempt_unsubscribe_of_principal(
-            query_count=15,
+            query_count=14,
             target_users=[self.example_user("cordelia")],
             is_realm_admin=True,
             is_subbed=True,
@@ -3112,7 +3112,7 @@ class StreamAdminTest(ZulipTestCase):
 
     def test_admin_remove_multiple_users_from_stream_legacy_emails(self) -> None:
         result = self.attempt_unsubscribe_of_principal(
-            query_count=17,
+            query_count=16,
             target_users=[self.example_user("cordelia"), self.example_user("prospero")],
             is_realm_admin=True,
             is_subbed=True,
@@ -3126,7 +3126,7 @@ class StreamAdminTest(ZulipTestCase):
 
     def test_remove_unsubbed_user_along_with_subbed(self) -> None:
         result = self.attempt_unsubscribe_of_principal(
-            query_count=14,
+            query_count=13,
             target_users=[self.example_user("cordelia"), self.example_user("iago")],
             is_realm_admin=True,
             is_subbed=True,
@@ -3143,7 +3143,7 @@ class StreamAdminTest(ZulipTestCase):
         fails gracefully.
         """
         result = self.attempt_unsubscribe_of_principal(
-            query_count=7,
+            query_count=6,
             target_users=[self.example_user("cordelia")],
             is_realm_admin=True,
             is_subbed=False,
@@ -3173,7 +3173,7 @@ class StreamAdminTest(ZulipTestCase):
         webhook_bot = self.example_user("webhook_bot")
         do_change_bot_owner(webhook_bot, bot_owner=other_user, acting_user=other_user)
         result = self.attempt_unsubscribe_of_principal(
-            query_count=8,
+            query_count=7,
             target_users=[webhook_bot],
             is_realm_admin=False,
             is_subbed=True,
@@ -3199,10 +3199,16 @@ class StreamAdminTest(ZulipTestCase):
         stream = self.make_stream("public_stream")
 
         def check_unsubscribing_user(
-            user: UserProfile, can_remove_subscribers_group: UserGroup, expect_fail: bool = False
+            user: UserProfile,
+            can_remove_subscribers_group: UserGroup,
+            expect_fail: bool = False,
+            stream_name_list: list[str] | None = None,
         ) -> None:
             self.login_user(user)
-            self.subscribe(cordelia, stream.name)
+            if stream_name_list is None:
+                stream_name_list = [stream.name]
+            for stream_name in stream_name_list:
+                self.subscribe(cordelia, stream_name)
             do_change_stream_group_based_setting(
                 stream,
                 "can_remove_subscribers_group",
@@ -3212,7 +3218,7 @@ class StreamAdminTest(ZulipTestCase):
             result = self.client_delete(
                 "/json/users/me/subscriptions",
                 {
-                    "subscriptions": orjson.dumps([stream.name]).decode(),
+                    "subscriptions": orjson.dumps(stream_name_list).decode(),
                     "principals": orjson.dumps([cordelia.id]).decode(),
                 },
             )
@@ -3221,13 +3227,14 @@ class StreamAdminTest(ZulipTestCase):
                 return
 
             json = self.assert_json_success(result)
-            self.assert_length(json["removed"], 1)
+            self.assert_length(json["removed"], len(stream_name_list))
             self.assert_length(json["not_removed"], 0)
 
         check_unsubscribing_user(self.example_user("hamlet"), leadership_group, expect_fail=True)
         check_unsubscribing_user(self.example_user("iago"), leadership_group)
-        # Owners can always unsubscribe others even when they are not a member
-        # allowed group.
+        # Owners can unsubscribe others when they are not a member of
+        # the allowed group since owners have the permission to
+        # administer all channels.
         check_unsubscribing_user(self.example_user("desdemona"), leadership_group)
 
         check_unsubscribing_user(self.example_user("othello"), managers_group, expect_fail=True)
@@ -3236,12 +3243,16 @@ class StreamAdminTest(ZulipTestCase):
 
         stream = self.make_stream("private_stream", invite_only=True)
         self.subscribe(self.example_user("hamlet"), stream.name)
-        # Non-admins are not allowed to unsubscribe others from private streams that they
-        # are not subscribed to even if they are member of the allowed group.
+        # Users are not allowed to unsubscribe others from streams they
+        # don't have access to even if they are a member of the allowed
+        # group. In this case, a non-admin who is not subscribed to the
+        # channel does not have access to the channel.
         check_unsubscribing_user(self.example_user("shiva"), leadership_group, expect_fail=True)
         check_unsubscribing_user(self.example_user("iago"), leadership_group)
-        # Owners can always unsubscribe others even when they are not a member
-        # allowed group.
+        # Users are allowed to unsubscribe others from private streams
+        # they have access to if they are a member of the allowed
+        # group. In this case, a user with the role `owner` is
+        # subscribed to the relevant channel.
         check_unsubscribing_user(self.example_user("desdemona"), leadership_group)
         self.subscribe(self.example_user("shiva"), stream.name)
         check_unsubscribing_user(self.example_user("shiva"), leadership_group)
@@ -3256,14 +3267,53 @@ class StreamAdminTest(ZulipTestCase):
         check_unsubscribing_user(self.example_user("iago"), setting_group)
         check_unsubscribing_user(self.example_user("shiva"), setting_group)
 
-        # Admins can unsubscribe others even when they are not a member of the
-        # allowed group.
+        # Owners can unsubscribe others when they are not a member of
+        # the allowed group since admins have the permission to
+        # administer all channels.
         setting_group = self.create_or_update_anonymous_group_for_setting(
             [hamlet],
             [],
         )
         check_unsubscribing_user(self.example_user("desdemona"), setting_group)
         check_unsubscribing_user(self.example_user("iago"), setting_group)
+
+        # A user who is part of can_administer_channel_group should be
+        # able to unsubscribe other users even if that user is not part
+        # of can_remove_subscribers_group. And even if that user is not
+        # subscribed to the channel in question.
+        othello = self.example_user("othello")
+        with self.assertRaises(Subscription.DoesNotExist):
+            get_subscription(stream.name, othello)
+        check_unsubscribing_user(othello, setting_group, expect_fail=True)
+        othello_group = self.create_or_update_anonymous_group_for_setting(
+            [othello],
+            [],
+        )
+        private_stream_2 = self.make_stream("private_stream_2")
+        do_change_stream_group_based_setting(
+            stream,
+            "can_administer_channel_group",
+            othello_group,
+            acting_user=None,
+        )
+        # If the user can only administer one of the channels, the test
+        # should fail.
+        check_unsubscribing_user(
+            othello,
+            setting_group,
+            expect_fail=True,
+            stream_name_list=[stream.name, private_stream_2.name],
+        )
+        # User can administer both channels now.
+        do_change_stream_group_based_setting(
+            private_stream_2,
+            "can_administer_channel_group",
+            othello_group,
+            acting_user=None,
+        )
+        check_unsubscribing_user(
+            othello, setting_group, stream_name_list=[stream.name, private_stream_2.name]
+        )
 
     def test_remove_invalid_user(self) -> None:
         """